+++ source/src/main/java/com/steinwurf/score/source/InvalidSnackPacketException.java
@@ -1,0 +1,7 @@
+package com.steinwurf.score.source;
+
+public class InvalidSnackPacketException extends Exception {
+    public InvalidSnackPacketException(String message) {
+        super(message);
+    }
+}